Move sor utils files into utils folder
